requirements.txt
APScheduler==3.6.3
astor==0.8.1
attrs==20.3.0
bandit==1.7.0
certifi==2020.12.5
chardet==4.0.0
darglint==1.7.0
docutils==0.16
environs==9.3.1
eradicate==2.0.0
flake8==3.8.4
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==20.11.1
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==4.0.0
flake8-eradicate==1.0.0
flake8-isort==4.0.0
flake8-polyfill==1.0.2
flake8-quotes==3.2.0
flake8-rst-docstrings==0.0.14
flake8-string-format==0.3.0
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
isort==5.7.0
marshmallow==3.10.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
pbr==5.5.1
pep8-naming==0.11.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
restructuredtext-lint==1.3.2
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
stevedore==3.3.0
testfixtures==6.17.1
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
